+++ aegisub/audio_provider_stream.h
@@ -68,7 +68,7 @@ public:
 	StreamAudioProvider();
 	~StreamAudioProvider();
 
-	void GetAudio(void *buf, long long start, long long count);
-	void Append(void *buf, long long count);
+	void GetAudio(void *buf, int64_t start, int64_t count);
+	void Append(void *buf, int64_t count);
 	void SetParams(int channels,int rate,int bps);
 };
